% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-fns_equal.R
\name{fns_equal}
\alias{fns_equal}
\title{Check whether two function objects have the same text definition.}
\usage{
fns_equal(x, y)
}
\arguments{
\item{x}{First function to compare.}

\item{y}{Second function to compare.}
}
\value{
Logical
}
\description{
Check whether two function objects have the same text definition.
}
