% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_read.R
\name{get_cache_type_list}
\alias{get_cache_type_list}
\title{Parses the `cache_type` and `_cache_fn` arguments for `cached_read` and returns a list containing:
 `write`: The write function.
 `read`: The read function.
 `ext`: The extension to use for the cache file.}
\usage{
get_cache_type_list(cache_type, write_cache_fn, read_cache_fn)
}
\arguments{
\item{cache_type}{The type of file to use for caching.
This can be one of two options:

 1. One of the following strings:
     "arrow" (same as `write_cache_fn=arrow::write_feather` and `read_cache_fn=arrow::read_feather`),
     "csv" (same as `readr::write_csv` and `readr::read_csv` if `readr` is installed; otherwise base R `utils::write.csv` and `utils::read.csv`)
 2. (Default) NULL:
     Uses `write_cache_fn` and `read_cache_fn` if provided. Otherwise, uses `"arrow"`, if installed, or `"csv"`.}

\item{write_cache_fn, read_cache_fn}{Functions used to write and read the cache file. If this is provided, the `cache_type` must be NULL.}
}
\description{
Parses the `cache_type` and `_cache_fn` arguments for `cached_read` and returns a list containing:
 `write`: The write function.
 `read`: The read function.
 `ext`: The extension to use for the cache file.
}
