% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_type.R
\name{get_csv_fns}
\alias{get_csv_fns}
\title{Get the CSV read/write function}
\usage{
get_csv_fns(type = NULL)
}
\arguments{
\item{type}{Type of csv read/write functions to get.
If \code{NULL}, returns the first installed.}
}
\value{
List of read/write functions.
}
\description{
Read functions are vectorized.
}
