% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_type.R
\name{get_csv_read_fn}
\alias{get_csv_read_fn}
\title{Get the first CSV Read function installed}
\usage{
get_csv_read_fn(read_type = NULL)
}
\arguments{
\item{read_type}{Type of csv read function to use. One of:
\itemize{
\item "readr": \code{readr::read_csv()}
\item "arrow": \code{vectorize_reader(arrow::read_csv_arrow)()}
\item "data.table": \code{vectorize_reader(data.table::fread)()}
\item "base": \code{vectorize_reader(utils::read.csv)()}
\item \code{NULL} (default): uses the first installed.
}}
}
\value{
Function that reads multiple paths to CSVs.
}
\description{
Get the first CSV Read function installed
}
