% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_read.R
\name{get_file_info}
\alias{get_file_info}
\title{Get information from fs::file_info that is expected to stay the same if the contents aren't modified.}
\usage{
get_file_info(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
For example, excludes access_time, which changes even if the contents are the same.
}
