% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_read.R
\name{parse_cache_type}
\alias{parse_cache_type}
\title{Parses the \code{cache_type} arguments for \code{cached_read} and returns a list containing:
\code{write}: The write function.
\code{read}: The read function.
\code{ext}: The extension to use for the cache file.}
\usage{
parse_cache_type(cache_type)
}
\arguments{
\item{cache_type}{The type of file to use for caching.
This can be one of two options:
\enumerate{
\item One of the following strings:
\itemize{
\item "arrow" to use \verb{arrow::<read|write>_feather},
\item "data.table" to use \verb{data.table::f<read|write>},
\item "csv" to use the first available of \verb{arrow::<read|write>_csv}, \code{readr}, \code{data.table}, or base R \code{utils}.
}
\item NULL:
\itemize{
\item (Default) Uses "arrow" if installed, otherwise "csv".
\item Uses \code{write_cache_fn}, \code{read_cache_fn}, and \code{cache_ext} if provided.
}
}}
}
\description{
Parses the \code{cache_type} arguments for \code{cached_read} and returns a list containing:
\code{write}: The write function.
\code{read}: The read function.
\code{ext}: The extension to use for the cache file.
}
