% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_cache.R
\name{with_cache}
\alias{with_cache}
\title{Cache via a file}
\usage{
with_cache(x, label, cache = NULL, type = NULL, force = FALSE)
}
\arguments{
\item{x}{The object to store in the cache. Must be compatible
with the cache type.}

\item{label}{A string to use as the name of the file to cache.}

\item{cache}{An existing \code{cachem} object or a path to an existing directory
to use for caching.}

\item{type}{A string describing the type of cache.
Must be \code{NULL} or one of 'rds', 'parquet', or 'csv'.
If \code{NULL} (default), uses 'rds'.}

\item{force}{If \code{TRUE}, forces evaluation even if the cache exists.}
}
\value{
The value of \code{x}.
}
\description{
If the cache exists, the object is retrieved from the cache.
Otherwise, it is evaluated and stored for subsequent retrieval.

Use \code{force=TRUE} to ensure the object is evaluated and stored
anew in the cache.

The object evaluated must be compatible with the cache type.
For example, a cache type of 'csv' or 'parquet' requires a
\code{data.frame} or similar type.
}
