% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_cache.R
\name{file_cache}
\alias{file_cache}
\title{Gets or creates a cachem object for use with other functions.}
\usage{
file_cache(cache = NULL, type = NULL)
}
\arguments{
\item{cache}{An existing cachem object or a path to an existing directory to use for caching.}

\item{type}{A string describing the type of cache.
Must be NULL or one of 'rds', 'parquet', or 'csv'.
If NULL (default), uses 'rds'.}
}
\value{
Either the cachem object provided, or a new \code{cachem::cache_disk} object.
}
\description{
Gets or creates a cachem object for use with other functions.
}
