% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-get_file_info.R
\name{get_file_info}
\alias{get_file_info}
\title{Extract File Information to Indicate if Contents Are Modified.}
\usage{
get_file_info(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Uses \code{file.info()} to get \code{size} and \code{mtime}.
}
