% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_reader.R
\name{vectorize_reader}
\alias{vectorize_reader}
\title{Vectorize a single-input read function to read multiple files}
\usage{
vectorize_reader(read_fn, file_path_to = NULL)
}
\arguments{
\item{read_fn}{The read function to vectorize. The first argument must be the
files to read.}

\item{file_path_to}{A string, which if provided, is the name of the column
containing the file paths in the result. See 'names_to' in
\code{\link[purrr:list_c]{purrr::list_rbind()}}.}
}
\value{
A version of \code{read_fn} that can read multiple paths.
}
\description{
The resulting vectorized read function still takes all the arguments of the
original function.

Uses \code{\link[purrr:list_c]{purrr::list_rbind()}} to bind the data frames, which generates
a data frame with a superset of the columns from all the files,
filling \code{NA} where data was not present.
}
\examples{
# Create multiple temporary paths to read.
tf <- tempfile()
dir.create(tf)

write.csv(mtcars, file.path(tf, "mtcars1.csv"))
write.csv(mtcars, file.path(tf, "mtcars2.csv"))
multiple_paths <- list.files(tf, full.names = TRUE)

try(read.csv(multiple_paths))
vectorize_reader(read.csv)(
  multiple_paths, header = FALSE
) |> dim()

try(arrow::read_csv_arrow(multiple_paths))
vectorize_reader(arrow::read_csv_arrow)(
  multiple_paths, col_names = FALSE
) |> dim()

try(data.table::fread(multiple_paths))
vectorize_reader(data.table::fread)(
  multiple_paths, header = FALSE
) |> dim()

unlink(tf)
}
\seealso{
\code{\link[purrr:list_c]{purrr::list_rbind()}}
}
