% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_reader.R
\name{vectorize_reader}
\alias{vectorize_reader}
\title{Vectorize a single-input read function to read multiple files at once.}
\usage{
vectorize_reader(read_fn, bind_fn = NULL, ...)
}
\arguments{
\item{read_fn}{The read function to vectorize.}

\item{bind_fn}{The function used to bind the results together.
If NULL (default) uses \code{dplyr::bind_rows}, if installed, or \code{rbind}.}

\item{...}{Arguments to \code{bind_fn}. For example, \code{dplyr::bind_rows} has a useful \code{.id} argument.}
}
\value{
A single data.frame or data.frame-like object.
}
\description{
Uses \code{dplyr::bind_rows} to bind the data frames, if installed. If not, uses \code{base::rbind}.
Note that these behave slightly differently:
\itemize{
\item \code{base::rbind} requires the data frames to have the same columns.
\item \code{dplyr::bind_rows} generates a data frame with all the columns, filling NA where data was not present.
}
}
\examples{
\dontrun{

path_list <- fs::dir_ls(DIR, glob="*.csv")

path_list |>
  vectorize_reader(read.csv)()

path_list |>
  vectorize_reader(arrow::read_csv_arrow)()

path_list |>
  vectorize_reader(data.table::fread)()

}
}
\seealso{
\code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}
}
